requirements.txt
python-gitlab~=4.5.0
numpy~=1.26.4